refactor: user service refactor
