--- api/src/services/user/handlers/addTeacher/options.js
@@ -1,18 +1,0 @@
-export const addTeacherOptions = {
-  schema: {
-    body: {
-      type: 'object',
-      properties: {
-        id: { type: 'number' },
-      },
-      required: ['id'],
-    },
-    response: {
-      '4xx': { $ref: '4xx#' },
-      '5xx': { $ref: '5xx#' },
-    },
-  },
-  async onRequest(req) {
-    await this.auth({ req, isAdminOnly: true });
-  },
-};
