--- api/src/services/user/handlers/deleteUser/index.js
@@ -1,4 +1,0 @@
-import { deleteUserHandler } from './handler';
-import { deleteUserOptions } from './options';
-
-export { deleteUserHandler, deleteUserOptions };
