--- api/src/services/user/handlers/deleteUser/options.js
@@ -1,18 +1,0 @@
-export const deleteUserOptions = {
-  schema: {
-    params: {
-      type: 'object',
-      properties: {
-        userId: { type: 'number' },
-      },
-      required: ['userId'],
-    },
-    response: {
-      '4xx': { $ref: '4xx#' },
-      '5xx': { $ref: '5xx#' },
-    },
-  },
-  async onRequest(req) {
-    await this.auth({ req, isAdminOnly: true });
-  },
-};
