--- api/src/services/user/handlers/getAllAuthors/handler.js
@@ -1,15 +1,0 @@
-export async function getAllAuthorsHandler({
-  query: { search, offset, limit },
-}) {
-  const {
-    models: { UserRole },
-  } = this;
-
-  const { total, results: authors } = await UserRole.getAllAuthors({
-    offset,
-    limit,
-    search,
-  });
-
-  return { total, authors };
-}
