--- api/src/services/user/handlers/getAllAuthors/index.js
@@ -1,4 +1,0 @@
-import { getAllAuthorsHandler } from './handler';
-import { getAllAuthorsOptions } from './options';
-
-export { getAllAuthorsHandler, getAllAuthorsOptions };
