--- api/src/services/user/handlers/getAllAuthors/options.js
@@ -1,12 +1,0 @@
-export const getAllAuthorsOptions = {
-  schema: {
-    querystring: { $ref: 'userSearch#' },
-    response: {
-      '4xx': { $ref: '4xx#' },
-      '5xx': { $ref: '5xx#' },
-    },
-  },
-  async onRequest(req) {
-    await this.auth({ req });
-  },
-};
