--- api/src/services/user/handlers/getAllUsers/handler.js
@@ -1,17 +1,0 @@
-export async function getAllUsersHandler({
-  user: { id: userId },
-  query: { search, offset, limit },
-}) {
-  const {
-    models: { User },
-  } = this;
-
-  const { total, results: data } = await User.getAllUsers({
-    userId,
-    offset,
-    limit,
-    search,
-  });
-
-  return { total, data };
-}
