--- api/src/services/user/handlers/getUser/index.js
@@ -1,4 +1,0 @@
-import { getUserHandler } from './handler';
-import { getUserOptions } from './options';
-
-export { getUserHandler, getUserOptions };
