--- api/src/services/user/handlers/removeTeacher/index.js
@@ -1,4 +1,0 @@
-import { removeTeacherHandler } from './handler';
-import { removeTeacherOptions } from './options';
-
-export { removeTeacherHandler, removeTeacherOptions };
