--- api/src/services/user/handlers/self/handler.js
@@ -1,9 +1,0 @@
-export async function selfHandler({ user: { id: userId } }) {
-  const {
-    models: { User },
-  } = this;
-
-  const user = await User.self({ userId });
-
-  return { ...user };
-}
