--- api/src/services/user/handlers/signIn/options.js
@@ -1,16 +1,0 @@
-export const signInOptions = {
-  schema: {
-    body: {
-      type: 'object',
-      properties: {
-        email: { type: 'string' },
-        password: { type: 'string' },
-      },
-      required: ['email', 'password'],
-    },
-    response: {
-      '4xx': { $ref: '4xx#' },
-      '5xx': { $ref: '5xx#' },
-    },
-  },
-};
