--- api/src/services/user/handlers/signUp/index.js
@@ -1,4 +1,0 @@
-import { signUpHandler } from './handler';
-import { signUpOptions } from './options';
-
-export { signUpHandler, signUpOptions };
