--- api/src/services/user/handlers/updateSelf/index.js
@@ -1,4 +1,0 @@
-import { updateSelfHandler } from './handler';
-import { updateSelfOptions } from './options';
-
-export { updateSelfOptions, updateSelfHandler };
