--- api/src/services/user/handlers/updateUser/index.js
@@ -1,4 +1,0 @@
-import { updateUserHandler } from './handler';
-import { updateUserOptions } from './options';
-
-export { updateUserHandler, updateUserOptions };
