--- api/src/services/user/handlers/updateUserLanguage/handler.js
@@ -1,15 +1,0 @@
-export async function updateUserLanguageHandler({
-  user: { id },
-  body: { language },
-}) {
-  const {
-    models: { User },
-  } = this;
-
-  const data = await User.updateLanguage({
-    userId: id,
-    language,
-  });
-
-  return { language: data?.language };
-}
