--- api/src/services/user/handlers/updateUserLanguage/index.js
@@ -1,4 +1,0 @@
-import { updateUserLanguageHandler } from './handler';
-import { updateUserLanguageOptions } from './options';
-
-export { updateUserLanguageHandler, updateUserLanguageOptions };
